LayoutTests/editing/selection/move-selection-forwards-at-beginning-of-table-by-sentence-granularity.html

<!DOCTYPE html>
<html>
<style>
table, th, td {
  border:1px solid black;
}
</style>
<body>
<div id="editor" contenteditable>
<table>
  <tr>
    <th  id="target">First</th>
    <th>Second</th> 
    <th>Third</th>
  </tr>
</table>
</div>
<script src="../../resources/dump-as-markup.js"></script>
<script>

Markup.description('This tests moving the selection forwards to the end of a table by sentence granularity. The caret should move to the end of the table.');

editor.focus();
getSelection().setPosition(target, 0);
getSelection().modify('move', 'right', 'sentence');

Markup.dump(editor);

</script>
</body>
</html>
